requirements.txt
anyio==3.6.2
boto3==1.26.18
botocore==1.29.18
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
motor==3.1.1
pydantic==1.10.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
